build: fixed some unboxing issues present to due a bug in the eclipse compiler for java 8. See Bug 440019

+++ plugins/de.cau.cs.kieler.klighd.piccolo/src/de/cau/cs/kieler/klighd/piccolo/internal/controller/DiagramZoomController.java
@@ -224,7 +224,7 @@ public class DiagramZoomController {
 
             while (parent != null && parent != displayedKNode.getParent()) {
                 final KShapeLayout parentLayout = parent.getData(KShapeLayout.class);
-                final double scale = parentLayout.getProperty(LayoutOptions.SCALE_FACTOR);
+                final double scale = parentLayout.getProperty(LayoutOptions.SCALE_FACTOR).doubleValue();
 
                 focusBounds.setSize(scale * focusBounds.width , scale * focusBounds.height);
                 focusBounds.setOrigin(scale * focusBounds.x, scale * focusBounds.y);
